django_requirements.txt
Django>=1.4.10,<1.7
django-jsonfield>=0.9.12
